Refactored packages - incorporated changes
