--- 1.1_declaring_and_printing_variables/com.thinkful/DeclareAndPrintSolution.java
@@ -1,24 +1,0 @@
-package com.thinkful;
-
-public class DeclareAndPrintSolution {
-
-  public static void main(String[] args) {
-
-    String highestTemp = "highest temperature in June in New York is 79°F";
-    System.out.println(highestTemp);
-
-    String courseName = "the name of this course is \"Java/Spring programming\"";
-    System.out.println(courseName);
-
-    String phoneNumber = "the phone number of the bus service is 555-5555";
-    System.out.println(phoneNumber);
-
-    String shoeCost = "the price of the shoe is $39.98";
-    System.out.println(shoeCost);
-
-    String jupiterMass = "the mass of the planet Jupiter is 1.89813 × 10²⁷ kilograms";
-    System.out.println(jupiterMass);
-
-  }
-
-}
