--- 1.2_simple_string_formatting/com.thinkful/SimpleStringFormattingSolution.java
@@ -1,41 +1,0 @@
-package com.thinkful;
-
-import java.util.Scanner;
-
-public class SimpleStringFormattingSolution {
-
-  public static void main(String[] args) {
-
-    String movieOne = "1. The Lord of the Rings: The Fellowship of the Rings";
-    System.out.println(movieOne);
-
-    String castOne = "Viggo Mortensen, Elijah Wood";
-    System.out.println(castOne);
-
-    String movieTwo = "2. Spider-Man Homecoming";
-    System.out.println(movieTwo);
-
-    String castTwo = "Tom Holland, Michael Keaton";
-    System.out.println(castTwo);
-
-    String movieThree = "3. Star Wars";
-    System.out.println(movieThree);
-
-    String castThree = "Carrie Fisher, Mark Hamill, Harrison Ford";
-    System.out.println(castThree);
-
-    String movieFour = "4. Indiana Jones and the Last Crusade";
-    System.out.println(movieFour);
-
-    String castFour = "Harrison Ford, Sean Connery";
-    System.out.println(castFour);
-
-    String movieFive = " 5. Lost in Translation";
-    System.out.println(movieFive);
-
-    String castFive = "Bill Murray, Scarlett Johansson";
-    System.out.println(castFive);
-
-  }
-
-}
